+++ src/ui/windows/TogglDesktop/TogglDesktop/TogglApi.cs
@@ -337,7 +337,7 @@ public         string ProjectAndTaskLabel;
 public         string StartTime;
 [MarshalAs(UnmanagedType.LPWStr)]
 public         string EndTime;
-public         uint DaysOfWeek;
+public         byte DaysOfWeek;
 public         IntPtr Next;
 
 public override string ToString()
@@ -1771,7 +1771,7 @@ private static extern Int64 toggl_autotracker_add_rule(
         string start_time,
         [MarshalAs(UnmanagedType.LPWStr)]
         string end_time,
-        uint days_of_week);
+        byte days_of_week);
 
 [DllImport(dll, CharSet = charset, CallingConvention = convention)]
 [return:MarshalAs(UnmanagedType.I1)]
@@ -1786,7 +1786,7 @@ private static extern bool toggl_autotracker_update_rule(
         string start_time,
         [MarshalAs(UnmanagedType.LPWStr)]
         string end_time,
-        uint days_of_week);
+        byte days_of_week);
 
 [DllImport(dll, CharSet = charset, CallingConvention = convention)]
 [return:MarshalAs(UnmanagedType.I1)]
